Bugfix! Fixed Darkmode toggle button.
