+++ carecrate/src/components/layout.tsx
@@ -4,12 +4,13 @@ import Navbar from './navbar';
 
 interface LayoutProps {
     children: React.ReactNode;
+    toggleMode: () => void;
 }
 
-export default function Layout({ children }: LayoutProps) {
+export default function Layout({ children, toggleMode }: LayoutProps) {
     return (
         <>
-            <Navbar />
+            <Navbar toggleMode={toggleMode}/>
             <Box component='main' sx={{ flex: '1 0 auto', minHeight: '100vh' }}>
                 { children }
             </Box>
